Move appendEdit from common.js to columnar.js
common.js is intended for code that is shared by frontend and backend,
but appendEdit is only needed in the backend.
